+++ src/main/java/com/mindlinksoft/recruitment/mychat/model/Conversation.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat;
+package com.mindlinksoft.recruitment.mychat.model;
 
 import java.util.Collection;
