+++ src/main/java/com/mindlinksoft/recruitment/mychat/model/Message.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat;
+package com.mindlinksoft.recruitment.mychat.model;
 
 import java.time.Instant;
