--- src/test/java/com/mindlinksoft/recruitment/mychat/ConversationExporterTests.java
@@ -1,73 +1,0 @@
-package com.mindlinksoft.recruitment.mychat;
-
-import com.google.gson.*;
-import com.google.gson.internal.LinkedTreeMap;
-import org.junit.Test;
-
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.lang.reflect.Type;
-import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Map;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Tests for the {@link ConversationExporter}.
- */
-public class ConversationExporterTests {
-    /**
-     * Tests that exporting a conversation will export the conversation correctly.
-     * @throws Exception When something bad happens.
-     */
-    @Test
-    public void testExportingConversationExportsConversation() throws IOException {
-        ConversationExporter exporter = new ConversationExporter();
-
-        exporter.exportConversation("chat.txt", "chat.json", null, "bob", new String[] {"hell"});
-
-        GsonBuilder builder = new GsonBuilder();
-        builder.registerTypeAdapter(Instant.class, new InstantDeserializer());
-
-        Gson g = builder.create();
-        JsonObject jo = g.fromJson(new InputStreamReader(new FileInputStream("chat.json")),JsonObject.class);
-        Conversation c = g.fromJson(jo.get("Conversation"), Conversation.class);
-
-        Message[] ms = new Message[c.messages.size()];
-        c.messages.toArray(ms);
-
-        assertEquals(ms[0].timestamp, Instant.ofEpochSecond(1448470901));
-        assertEquals(ms[0].senderId, "bob");
-        assertEquals(ms[0].content, "Hello there!");
-
-        assertEquals(ms[1].timestamp, Instant.ofEpochSecond(1448470906));
-        assertEquals(ms[1].senderId, "bob");
-        assertEquals(ms[1].content, "I'm good thanks, do you like pie?");
-
-        assertEquals(ms[2].timestamp, Instant.ofEpochSecond(1448470914));
-        assertEquals(ms[2].senderId, "bob");
-        assertEquals(ms[2].content, "No, just want to know if there's anybody else in the pie society...");
-
-    }
-
-    class InstantDeserializer implements JsonDeserializer<Instant> {
-
-        @Override
-        public Instant deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
-            if (!jsonElement.isJsonPrimitive()) {
-                throw new JsonParseException("Expected instant represented as JSON number, but no primitive found.");
-            }
-
-            JsonPrimitive jsonPrimitive = jsonElement.getAsJsonPrimitive();
-
-            if (!jsonPrimitive.isNumber()) {
-                throw new JsonParseException("Expected instant represented as JSON number, but different primitive found.");
-            }
-
-            return Instant.ofEpochSecond(jsonPrimitive.getAsLong());
-        }
-    }
-}
